==> 2 - preprocessing.ipynb <==
import pandas as pd

df = pd.read_excel("data/bronze/feedbacks-2023.xlsx")

df['quarter'] = pd.to_datetime(df['date']).dt.year.astype(str) + 'Q' + pd.to_datetime(df['date']).dt.quarter.astype(str)

categories = df["feedback_category"].unique().tolist()
print(f"Categories: {', '.join(categories)}")
sentiments = df["sentiment"].unique().tolist()
print(f"Sentiments: {', '.join(sentiments)}")

cleansed_cols = ["feedback_id", "customer_id", "date", "quarter", "feedback_text"]
text_deduped_cols = ["feedback_text"]

cleansed_df = df[cleansed_cols]
text_deduped_df = df[text_deduped_cols].drop_duplicates("feedback_text").reset_index(drop=True)

cleansed_df.sample(5, random_state=42)

text_deduped_df

cleansed_df.to_parquet("data/silver/cleansed-feedbacks-2023.parquet", index=False)
text_deduped_df.to_parquet("data/silver/feedback-text-deduped-2023.parquet", index=False)
