==> 3 - opinion_mining.ipynb <==
import json
import dotenv
import pandas as pd
from langchain_core.prompts import ChatPromptTemplate
from langchain_openai import ChatOpenAI
from pydantic import BaseModel, Field
from devtools import pprint

dotenv.load_dotenv(dotenv_path=".env")

feedbacks_deduped = pd.read_parquet("data/silver/feedback-text-deduped-2023.parquet")

feedbacks_deduped

### Aspect-Based Sentiment Analysis with Category and Opinion Extraction

sentiments = ["positive", "negative", "neutral"]
categories = ["Product Quality", "Delivery", "Pricing", "Customer Service", "Website Usability", "Other"]

extraction_prompt = ChatPromptTemplate.from_template(
    """
Extract the desired information from the following customer feedback.

Only extract the properties you were instructed to in the 'AspectBasedSentimentAnalysisWithCategoryAndOpinionExtraction' function.

Feedback:
{input}
"""
)


class AspectSentimentWithCategoryAndOpinion(BaseModel):
    aspect: str = Field(
        ...,
        description="A mentioned aspect (explicitly)"
    )
    opinion: str = Field(
        ...,
        description="The opinion related to the extracted aspect (explicitly)"
    )
    sentiment: str = Field(
        ...,
        description="The sentiment related to the extracted aspect",
        enum=sentiments
    )
    category: str = Field(
        ...,
        description="The category related to the extracted aspect",
        enum=categories
    )


class AspectBasedSentimentAnalysisWithCategoryAndOpinionExtraction(BaseModel):
    aspect_based_sentiments_with_categories: list[AspectSentimentWithCategoryAndOpinion] = Field(
        ...,
        description="A list of 'AspectSentimentWithCategoryAndOpinion' where each contains the sentiment, category and opinion related to a mentioned aspect"
    )

gpt4o_extraction_llm = ChatOpenAI(temperature=0, model="gpt-4o").with_structured_output(
    AspectBasedSentimentAnalysisWithCategoryAndOpinionExtraction
)
gpt4omini_extraction_llm = ChatOpenAI(temperature=0, model="gpt-4o-mini").with_structured_output(
    AspectBasedSentimentAnalysisWithCategoryAndOpinionExtraction
)
gpt35turbo_extraction_llm = ChatOpenAI(temperature=0, model="gpt-3.5-turbo").with_structured_output(
    AspectBasedSentimentAnalysisWithCategoryAndOpinionExtraction
)

gpt4o_extraction_chain = extraction_prompt | gpt4o_extraction_llm
gpt4omini_extraction_chain = extraction_prompt | gpt4omini_extraction_llm
gpt35turbo_extraction_chain = extraction_prompt | gpt35turbo_extraction_llm

def majority_voting_classifier(chain1, chain2, chain3, input):
    answ1 = chain1.invoke(input)
    answ2 = chain2.invoke(input)
    answ3 = chain3.invoke(input)
    answers_list = [
        answ1.json().lower(),
        answ2.json().lower(),
        answ3.json().lower()
    ]
    answers_set = set(answers_list)
    answer = max(answers_set, key=answers_list.count)
    votes = answers_list.count(answer)
    agreement = votes/3
    return json.loads(answer)["aspect_based_sentiments_with_categories"], votes, agreement


answer, votes, agreement = majority_voting_classifier(
    chain1=gpt4o_extraction_chain,
    chain2=gpt4omini_extraction_chain,
    chain3=gpt35turbo_extraction_chain,
    input="The delivery was fast and the product is good.",
)

print(f"Number of votes on final answer: {votes} ({int(100 * agreement)}% agreement)")
pprint(answer)

feedbacks_deduped[["extraction", "votes", "agreement"]] = feedbacks_deduped.apply(
    lambda row: majority_voting_classifier(
        chain1=gpt4o_extraction_chain,
        chain2=gpt4omini_extraction_chain,
        chain3=gpt35turbo_extraction_chain,
        input=row["feedback_text"],
    ),
    axis=1,
    result_type="expand"
)

feedbacks_deduped

print(f"average LLM agreement: {round(100 * feedbacks_deduped['agreement'].mean(), 2)}%")

def debug_classifier(chain1, chain2, chain3, input):
    answ1 = chain1.invoke(input)
    answ2 = chain2.invoke(input)
    answ3 = chain3.invoke(input)
    pprint(answ1.aspect_based_sentiments_with_categories)
    pprint(answ2.aspect_based_sentiments_with_categories)
    pprint(answ3.aspect_based_sentiments_with_categories)
    
debug_classifier(
    chain1=gpt4o_extraction_chain,
    chain2=gpt4omini_extraction_chain,
    chain3=gpt35turbo_extraction_chain,
    input="Easy to use website and quick checkout.",
)

debug_classifier(
    chain1=gpt4o_extraction_chain,
    chain2=gpt4omini_extraction_chain,
    chain3=gpt35turbo_extraction_chain,
    input="I had issues with the website.",
)

feedbacks_silver = pd.read_parquet("data/silver/cleansed-feedbacks-2023.parquet")
granular_opinions = feedbacks_silver.merge(feedbacks_deduped, on="feedback_text", how="inner")
granular_opinions = granular_opinions.drop(columns="feedback_text").explode("extraction").reset_index(drop=True)
granular_opinions = pd.concat(
    [granular_opinions.drop(columns=['extraction']), pd.json_normalize(granular_opinions['extraction'])],
    axis=1
)

granular_opinions

granular_opinions.to_parquet("data/gold/granular-opinions-2023.parquet", index=False)
